Move scheme to pkg, fix imports, remove unnecessary AddToScheme calls
