+++ pkg/apis/wksprovider/machine/crds/crds_vfsdata.go
@@ -26,9 +26,9 @@ var CRDs = func() http.FileSystem {
 		"/cluster.weave.works_existinginfraclusters.yaml": &vfsgen۰CompressedFileInfo{
 			name:             "cluster.weave.works_existinginfraclusters.yaml",
 			modTime:          time.Date(2020, 1, 1, 0, 0, 0, 0, time.UTC),
-			uncompressedSize: 5369,
+			uncompressedSize: 5346,
 
-			compressedContent: []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xdc\x58\xcd\x8e\x1b\x37\x0c\xbe\xcf\x53\x10\xe9\x21\x97\x7a\x9c\x20\x97\xc2\xb7\xc5\x36\x05\x82\xa4\xc1\x22\xbb\x4d\x0f\x45\x0f\xb4\x44\xcf\xa8\xd6\x48\x2a\x25\x39\xbb\x79\xfa\x42\x9a\x19\xdb\xbb\xf3\x63\x7b\x17\xed\xa1\x3a\x59\x34\x49\x7d\xfa\x48\x51\xd4\x14\x8b\xc5\xa2\x40\xa7\xbe\x12\x7b\x65\xcd\x0a\xd0\x29\xba\x0f\x64\xd2\xcc\x97\xdb\x9f\x7c\xa9\xec\x72\xf7\x76\x4d\x01\xdf\x16\x5b\x65\xe4\x0a\xae\xa3\x0f\xb6\xf9\x42\xde\x46\x16\xf4\x33\x6d\x94\x51\x41\x59\x53\x34\x14\x50\x62\xc0\x55\x01\x80\xc6\xd8\x80\x49\xec\xd3\x14\x40\x58\x13\xd8\x6a\x4d\xbc\xa8\xc8\x94\xdb\xb8\xa6\x75\x54\x5a\x12\xe7\x15\xfa\xf5\x77\x6f\xca\x77\xe5\x9b\x02\x40\x30\x65\xf3\x3b\xd5\x90\x0f\xd8\xb8\x15\x98\xa8\x75\x01\x60\xb0\xa1\x15\xd0\xbd\xf2\x41\x99\x4a\x99\x0d\xa3\xd0\xd1\x07\x62\x5f\x76\x3f\xca\x6f\x84\x3b\x2a\xbf\x59\xde\xfa\x02\x40\xe3\x9a\x74\x8f\xa3\xd3\xb8\x5f\xec\x37\x87\xda\xd5\xf8\x6e\x05\xfd\xaf\xc2\x3b\x12\x49\xbb\x62\x1b\xdd\x0a\xc6\x9d\x26\x18\x9d\xcf\x96\x97\xf7\x1d\xa2\x0f\x09\xd1\x75\x6b\x93\xff\xd6\xca\x87\x8f\x93\x2a\x9f\x94\x0f\x59\xcd\xe9\xc8\xa8\x27\x76\x96\x35\xbc\x32\x55\xd4\xc8\xe3\x3a\x05\x80\x17\xd6\xd1\x0a\x3e\x27\x68\x0e\x05\xc9\x02\x60\x87\x5a\xc9\x4c\x65\x0b\xd6\x3a\x32\x57\x37\x1f\xbe\xbe\xbb\x15\x35\x35\xd8\x0a\x01\x1c\x5b\x47\x1c\x54\xbf\xa7\x34\x8e\x12\x63\x2f\x03\x90\xe4\x05\x2b\x97\x3d\xc2\xeb\xe4\xaa\xd5\x01\x99\x52\x81\x3c\x84\x9a\xa0\x0b\x28\x49\xf0\x79\x19\xb0\x1b\x08\xb5\xf2\xc0\xe4\x98\x3c\x99\x36\x39\x8e\xdc\x42\x52\x41\x03\x76\xfd\x17\x89\x50\xc2\x2d\x71\x72\x02\xbe\xb6\x51\xcb\x94\x3f\x3b\xe2\x00\x4c\xc2\x56\x46\x7d\xdf\x7b\xf6\x10\x6c\x5e\x52\x63\xa0\x8e\xca\x7e\x28\x13\x88\x0d\xea\x44\x42\xa4\x1f\x01\x8d\x84\x06\x1f\x80\x29\xad\x01\xd1\x1c\x79\xcb\x2a\xbe\x84\x5f\x2d\x13\x28\xb3\xb1\x2b\xa8\x43\x70\x7e\xb5\x5c\x56\x2a\xf4\x47\x41\xd8\xa6\x89\x46\x85\x87\x65\x4e\x68\xb5\x8e\xc1\xb2\x5f\x4a\xda\x91\x5e\x7a\x55\x2d\x90\x45\xad\x02\x89\x10\x99\x96\xe8\xd4\x22\x03\x37\xf9\x24\x94\x8d\xfc\x81\xbb\x73\xe3\x5f\x1f\x21\x0d\x0f\x29\x6c\x3e\xb0\x32\xd5\x5e\x9c\xb3\x6a\x92\xf7\x94\x50\xa0\x3c\x60\x67\xd6\xe2\x3f\xd0\x9b\x44\x89\x95\x2f\xef\x6f\xef\xa0\x5f\x34\x87\xe0\x31\xe7\x99\xed\x83\x99\x3f\x10\x9f\x88\x52\x66\x43\xdc\x06\x6e\xc3\xb6\xc9\x1e\xc9\x48\x67\x95\x09\x79\x22\xb4\x22\xf3\x98\x74\x1f\xd7\x8d\x0a\x29\xd2\x7f\x47\xf2\x21\xc5\xa7\x84\xeb\x5c\x10\x60\x4d\x10\x9d\xc4\x40\xb2\x84\x0f\x06\xae\xb1\x21\x7d\x8d\x9e\xfe\x75\xda\x13\xc3\x7e\x91\x28\x3d\x4d\xfc\x71\x1d\x7b\xac\xd8\xb2\xb5\x17\xf7\x75\xa2\x1f\x63\x67\x28\x9f\x23\x29\xf7\x75\xf0\x28\x39\x03\x35\x03\xe1\x93\x30\x5f\x25\xcb\x4e\xb4\x26\x9f\xce\x47\x76\x96\x72\x5e\x19\x1f\x50\x6b\x48\xb3\x1c\x89\xb6\x4a\x0d\x1c\x4e\xa1\xea\x97\x73\xa3\xf2\x49\x7c\x33\xc4\x0d\x15\x90\x19\x1f\x46\xfe\xcf\x35\x7c\xc6\x6e\xd2\xb1\x43\xc6\x29\x54\x28\x65\xbe\x88\x50\xdf\xcc\x6e\xf9\x02\xfc\x4f\x02\xde\x8f\x94\xda\x8a\x49\x0e\xbd\x2f\xf2\xd6\x06\xe2\x19\x6f\x53\x44\xa1\x53\xed\x49\x7c\xba\xc8\x5c\x3c\x0f\x14\xdc\x5e\x7d\x1e\xdd\xfc\x4c\x54\x4f\x70\x32\x17\x51\xba\x0f\x8c\x57\x5c\xc5\x26\xd5\x90\x0b\xd7\x9d\x4f\x50\x98\x4d\x97\x33\x70\xa7\x91\x8b\xe3\x0b\x3c\x4c\x07\x1c\xa6\x83\xde\xfe\x95\x97\xbe\x3c\xc5\xe6\xf8\x9e\x34\xc4\x18\xea\x54\xf4\x44\xbe\x5c\x7f\xa7\x75\x6d\xed\xf6\x92\xfc\x11\x28\x6a\xba\xbb\xfb\x34\xb6\xcf\x59\x92\x3c\x09\xa6\xf0\x8b\xd2\xa3\x2c\xcf\x9a\x46\xd6\x17\xda\x4c\x45\x63\x71\x04\x63\xf0\x57\x64\x7d\x09\x8d\x96\xd5\xf7\x17\xb1\x78\xd5\x39\x21\xf9\x1c\x3e\xb3\x8b\xdf\x0c\xbe\xc8\xc9\xff\x2a\x28\x42\xdb\x28\x6f\xd8\xee\x94\x1c\xd6\xc4\x49\x5c\xdd\xb3\xe3\x46\xa3\xa1\xf7\x5d\xe3\x72\xbe\x31\xab\x4b\xc2\xfe\xb4\x61\x3b\xe9\x3f\x3b\x44\xb1\xc5\xea\xf2\x00\xed\x86\x8d\xf9\x19\x76\xd3\x41\x4a\xe0\x07\xc2\x0e\xdc\x40\xde\x2d\x7e\x6e\xe8\x52\x43\x77\xc3\xf6\xfe\xe1\x6c\xe6\x55\x83\x15\x7d\x21\x67\xbd\x0a\x96\xcf\xb7\x4b\xcf\x4a\x4d\x61\xf2\x26\x9a\xb8\x85\xe6\x6f\xa0\x67\x37\x2b\x33\xb7\xce\xac\xe5\xc5\xed\xc5\xd4\x2d\xf3\x8c\xb6\xc3\x0e\x28\x98\x23\x67\xa3\xf4\x38\x67\x2f\xba\xee\x25\xa5\xc7\xed\xd1\x9b\x75\x6c\x9c\x71\xeb\xb7\x8f\x9e\x69\x17\xa7\x91\x40\x5b\x42\x36\xaa\x6a\xd0\xcd\x29\x9d\x85\x27\x8d\x2d\x0d\xb2\xf9\x19\x7e\xe6\xfb\x11\xc8\x29\xb1\xc7\x3d\xa3\xb3\xa5\xb1\xee\xfc\x18\xc8\x64\x7b\x72\xba\x29\x3a\x0a\xe3\x84\x46\x1b\xa0\xff\xa8\x39\xf2\xbe\xfe\x48\x0f\x37\x18\xea\xb3\xcb\x49\xf4\x67\x5f\x35\x63\x64\x2c\xd2\x0d\xf2\x68\x7e\xc0\xf0\x48\x9c\xd6\x29\x4e\x6c\xc1\x07\x0c\xd1\x9f\xf3\xdc\x64\x42\x39\x51\x32\xd7\xd6\x6a\x42\x73\x12\x77\x76\x31\x8f\xe8\x89\xe8\xf0\x05\xaf\xff\x8a\xb6\x97\x65\x78\x8b\xee\xbb\xdd\xd1\xdf\xa9\x41\xe1\x1d\xc9\x15\x04\xee\x8a\x97\x0f\x96\xd3\x6d\xd8\x4a\x0e\x5b\x46\x21\xc8\x05\x92\x9f\x9f\x7e\x74\x7b\xf5\x2a\x4f\xfa\x6f\x67\x79\x2a\xac\x69\x5f\x41\x7e\x05\x7f\xfc\x59\xb4\x5e\x49\x7e\xed\xd1\x24\xe1\x3f\x01\x00\x00\xff\xff\x27\x55\x96\xd0\xf9\x14\x00\x00"),
+			compressedContent: []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xdc\x58\xcd\x8e\x1b\x37\x0c\xbe\xcf\x53\x10\xe9\x21\x97\x7a\x9c\x20\x97\xc2\xb7\xc5\x36\x05\x82\xa4\xc1\x22\xbb\x4d\x0f\x45\x0f\xb4\x44\xcf\xa8\xd6\x48\x2a\x25\x39\xbb\x79\xfa\x42\x9a\x19\xaf\xd7\xf3\x63\x7b\x17\xed\xa1\x3a\x59\x34\x49\x7d\xfa\x48\x51\xd4\x14\x8b\xc5\xa2\x40\xa7\xbe\x12\x7b\x65\xcd\x0a\xd0\x29\xba\x0f\x64\xd2\xcc\x97\xdb\x9f\x7c\xa9\xec\x72\xf7\x76\x4d\x01\xdf\x16\x5b\x65\xe4\x0a\xae\xa3\x0f\xb6\xf9\x42\xde\x46\x16\xf4\x33\x6d\x94\x51\x41\x59\x53\x34\x14\x50\x62\xc0\x55\x01\x80\xc6\xd8\x80\x49\xec\xd3\x14\x40\x58\x13\xd8\x6a\x4d\xbc\xa8\xc8\x94\xdb\xb8\xa6\x75\x54\x5a\x12\xe7\x15\xfa\xf5\x77\x6f\xca\x77\xe5\x9b\x02\x40\x30\x65\xf3\x3b\xd5\x90\x0f\xd8\xb8\x15\x98\xa8\x75\x01\x60\xb0\xa1\x15\xd0\xbd\xf2\x41\x99\x4a\x99\x0d\xa3\xd0\xd1\x07\x62\x5f\x76\x3f\xca\x6f\x84\x3b\x2a\xbf\x59\xde\xfa\x02\x40\xe3\x9a\x74\x8f\xa3\xd3\xb8\x5f\xec\x37\x87\xda\xd5\xf8\x6e\x05\xfd\xaf\xc2\x3b\x12\x49\xbb\x62\x1b\xdd\x0a\xc6\x9d\x26\x18\x9d\xcf\x96\x97\xf7\x1d\xa2\x0f\x09\xd1\x75\x6b\x93\xff\xd6\xca\x87\x8f\x93\x2a\x9f\x94\x0f\x59\xcd\xe9\xc8\xa8\x27\x76\x96\x35\xbc\x32\x55\xd4\xc8\xe3\x3a\x05\x80\x17\xd6\xd1\x0a\x3e\x27\x68\x0e\x05\xc9\x02\x60\x87\x5a\xc9\x4c\x65\x0b\xd6\x3a\x32\x57\x37\x1f\xbe\xbe\xbb\x15\x35\x35\xd8\x0a\x01\x1c\x5b\x47\x1c\x54\xbf\xa7\x34\x0e\x12\x63\x2f\x03\x90\xe4\x05\x2b\x97\x3d\xc2\xeb\xe4\xaa\xd5\x01\x99\x52\x81\x3c\x84\x9a\xa0\x0b\x28\x49\xf0\x79\x19\xb0\x1b\x08\xb5\xf2\xc0\xe4\x98\x3c\x99\x36\x39\x0e\xdc\x42\x52\x41\x03\x76\xfd\x17\x89\x50\xc2\x2d\x71\x72\x02\xbe\xb6\x51\xcb\x94\x3f\x3b\xe2\x00\x4c\xc2\x56\x46\x7d\xdf\x7b\xf6\x10\x6c\x5e\x52\x63\xa0\x8e\xca\x7e\x28\x13\x88\x0d\xea\x44\x42\xa4\x1f\x01\x8d\x84\x06\x1f\x80\x29\xad\x01\xd1\x1c\x78\xcb\x2a\xbe\x84\x5f\x2d\x13\x28\xb3\xb1\x2b\xa8\x43\x70\x7e\xb5\x5c\x56\x2a\xf4\x47\x41\xd8\xa6\x89\x46\x85\x87\x65\x4e\x68\xb5\x8e\xc1\xb2\x5f\x4a\xda\x91\x5e\x7a\x55\x2d\x90\x45\xad\x02\x89\x10\x99\x96\xe8\xd4\x22\x03\x37\xf9\x24\x94\x8d\xfc\x81\xbb\x73\xe3\x5f\x1f\x20\x0d\x0f\x29\x6c\x3e\xb0\x32\xd5\x5e\x9c\xb3\x6a\x92\xf7\x94\x50\xa0\x3c\x60\x67\xd6\xe2\x7f\xa4\x37\x89\x12\x2b\x5f\xde\xdf\xde\x41\xbf\x68\x0e\xc1\x53\xce\x33\xdb\x8f\x66\xfe\x91\xf8\x44\x94\x32\x1b\xe2\x36\x70\x1b\xb6\x4d\xf6\x48\x46\x3a\xab\x4c\xc8\x13\xa1\x15\x99\xa7\xa4\xfb\xb8\x6e\x54\x48\x91\xfe\x3b\x92\x0f\x29\x3e\x25\x5c\xe7\x82\x00\x6b\x82\xe8\x24\x06\x92\x25\x7c\x30\x70\x8d\x0d\xe9\x6b\xf4\xf4\xaf\xd3\x9e\x18\xf6\x8b\x44\xe9\x69\xe2\x0f\xeb\xd8\x53\xc5\x96\xad\xbd\xb8\xaf\x13\xfd\x18\x3b\x43\xf9\x1c\x49\xb9\xaf\x83\x07\xc9\x19\xa8\x19\x08\x8f\xc2\x7c\x95\x2c\x3b\xd1\x9a\x7c\x3a\x1f\xd9\x59\xca\x79\x65\x7c\x40\xad\x21\xcd\x72\x24\xda\x2a\x35\x70\x38\x85\xaa\x5f\xce\x8d\xca\x27\xf1\xcd\x10\x37\x54\x40\x66\x7c\x18\xf9\x3f\xd7\xf0\x19\xbb\x49\xc7\x0e\x19\xa7\x50\xa1\x94\xf9\x22\x42\x7d\x33\xbb\xe5\x0b\xf0\x1f\x05\xbc\x1f\x29\xb5\x15\x93\x1c\x7a\x5f\xe4\xad\x0d\xc4\x33\xde\xa6\x88\x42\xa7\xda\x93\x78\xbc\xc8\x5c\x3c\x1f\x29\xb8\xbd\xfa\x3c\xba\xf9\x99\xa8\x9e\xe0\x64\x2e\xa2\x74\x1f\x18\xaf\xb8\x8a\x4d\xaa\x21\x17\xae\x3b\x9f\xa0\x30\x9b\x2e\x67\xe0\x4e\x23\x17\xc7\x17\x78\x98\x0e\x38\x4c\x07\xbd\xfd\x2b\x2f\x7d\x79\x8a\xcd\xf1\x3d\x69\x88\x31\xd4\xa9\xe8\x89\x7c\xb9\xfe\x4e\xeb\xda\xda\xed\x25\xf9\x23\x50\xd4\x74\x77\xf7\x69\x6c\x9f\xb3\x24\x79\x12\x4c\xe1\x17\xa5\x47\x59\x9e\x35\x8d\xac\x2f\xb4\x99\x8a\xc6\xe2\x00\xc6\xe0\xaf\xc8\xfa\x12\x1a\x2d\xab\xef\x2f\x62\xf1\xaa\x73\x42\xf2\x39\x7c\x66\x17\xbf\x19\x7c\x91\x93\xff\x55\x50\x84\xb6\x51\xde\xb0\xdd\x29\x39\xac\x89\x93\xb8\xba\x67\xc7\x8d\x46\x43\xef\xbb\xc6\xe5\x7c\x63\x56\x97\x84\xfd\xb8\x61\x3b\xe9\x3f\x3b\x44\xb1\xc5\xea\xf2\x00\xed\x86\x8d\xf9\x19\x76\xd3\x41\x4a\xe0\x07\xc2\x0e\xdc\x40\xde\x2d\x7e\x6e\xe8\x52\x43\x77\xc3\xf6\xfe\xe1\x6c\xe6\x55\x83\x15\x7d\x21\x67\xbd\x0a\x96\xcf\xb7\x4b\xcf\x4a\x4d\x61\xf2\x26\x9a\xb8\x85\xe6\x6f\xa0\x67\x37\x2b\x33\xb7\xce\xac\xe5\xc5\xed\xc5\xd4\x2d\xf3\x8c\xb6\xc3\x0e\x28\x98\x23\x67\xa3\xf4\x38\x67\x2f\xba\xee\x25\xa5\xc7\xed\xc1\x9b\x75\x6c\x9c\x71\xeb\xb7\x8f\x9e\x69\x17\xa7\x91\x40\x5b\x42\x36\xaa\x6a\xd0\xcd\x29\x9d\x85\x27\x8d\x2d\x0d\xb2\xf9\x19\x7e\xe6\xfb\x11\xc8\x29\xb1\xc7\x3d\xa3\xb3\xa5\xb1\xee\xfc\x10\xc8\x64\x7b\x72\xba\x29\x3a\x08\xe3\x84\x46\x1b\xa0\xff\xa8\x39\xf2\xbe\xfe\x48\x0f\x37\x18\xea\xb3\xcb\x49\xf4\x67\x5f\x35\x63\x64\x2c\xd2\x0d\xf2\x64\x9e\x1c\x16\x27\xb0\xfa\x80\x21\xfa\x73\xde\x95\x4c\x28\x27\x6a\xe3\xda\x5a\x4d\x68\x4e\x02\xcc\x2e\xe6\x11\x1d\x89\x1e\x3f\xd5\xf5\x9f\xcb\xf6\xb2\x0c\x6f\xd1\x7d\xa0\x3b\xf8\x3b\x75\x22\xbc\x23\xb9\x82\xc0\x5d\x95\xf2\xc1\x72\xba\xf6\x5a\xc9\xe3\x96\x51\x08\x72\x81\xe4\xe7\xe3\xaf\x6b\xaf\x5e\xe5\x49\xff\x91\x2c\x4f\x85\x35\xed\x73\xc7\xaf\xe0\x8f\x3f\x8b\xd6\x2b\xc9\xaf\x3d\x9a\x24\xfc\x27\x00\x00\xff\xff\xd1\x42\xf0\xb3\xe2\x14\x00\x00"),
 		},
 		"/cluster.weave.works_existinginframachiness.yaml": &vfsgen۰CompressedFileInfo{
 			name:             "cluster.weave.works_existinginframachiness.yaml",
